requirements-monitoring.txt
# Monitoring and Observability Dependencies

# Prometheus Client
prometheus-client==0.19.0

# System Metrics
psutil==5.9.6

# OpenTelemetry Core
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# OpenTelemetry Exporters
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlite3==0.42b0

# Optional: Additional exporters
# opentelemetry-exporter-prometheus==0.42b0
# opentelemetry-exporter-otlp==1.21.0
